BZ-1022910 - Retrieving task instances when using PostgreSQL (Large Object facility requires transacitons and eager initialization) (cherry picked from commit df78e14ec901efa1f15acf38e1871f725179f25e)

+++ kogito-jbpm/jbpm-human-task/jbpm-human-task-core/src/test/java/org/jbpm/services/task/HumanTaskServicesBaseTest.java
@@ -15,9 +15,12 @@
  */
 package org.jbpm.services.task;
 
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.io.StringWriter;
@@ -28,11 +31,13 @@ import java.util.Arrays;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Properties;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.Marshaller;
 import javax.xml.bind.Unmarshaller;
 
+import org.jbpm.persistence.util.PersistenceUtil;
 import org.jbpm.services.task.impl.model.xml.JaxbContent;
 import org.jbpm.services.task.utils.ContentMarshallerHelper;
 import org.jbpm.services.task.utils.MVELUtils;
@@ -52,8 +57,10 @@ public abstract class HumanTaskServicesBaseTest {
 
 
     public void tearDown() {
-        int removeAllTasks = taskService.removeAllTasks();
-        logger.debug("Number of tasks removed {}", removeAllTasks);
+        if( taskService != null ) { 
+            int removeAllTasks = taskService.removeAllTasks();
+            logger.debug("Number of tasks removed {}", removeAllTasks);
+        }
     }
 
     @SuppressWarnings({"unchecked", "rawtypes"})
@@ -184,17 +191,89 @@ public abstract class HumanTaskServicesBaseTest {
         Assert.assertEquals(orig, roundTrip);
     }
     
+    protected static final String DATASOURCE_PROPERTIES = "/datasource.properties";
+    
+    protected static final String MAX_POOL_SIZE = "maxPoolSize";
+    protected static final String ALLOW_LOCAL_TXS = "allowLocalTransactions";
+    
+    protected static final String DATASOURCE_CLASS_NAME = "className";
+    protected static final String DRIVER_CLASS_NAME = "driverClassName";
+    protected static final String USER = "user";
+    protected static final String PASSWORD = "password";
+    protected static final String JDBC_URL = "url";
+    
     protected PoolingDataSource setupPoolingDataSource() {
-        PoolingDataSource pds = new PoolingDataSource();
-        pds.setUniqueName("jdbc/jbpm-ds");
-        pds.setClassName("bitronix.tm.resource.jdbc.lrc.LrcXADataSource");
-        pds.setMaxPoolSize(5);
-        pds.setAllowLocalTransactions(true);
-        pds.getDriverProperties().put("user", "sa");
-        pds.getDriverProperties().put("password", "");
-        pds.getDriverProperties().put("url", "jdbc:h2:mem:jbpm-db;MVCC=true");
-        pds.getDriverProperties().put("driverClassName", "org.h2.Driver");
+        Properties dsProps = getDatasourceProperties();
+        PoolingDataSource pds = PersistenceUtil.setupPoolingDataSource(dsProps, "jdbc/jbpm-ds", false);
         pds.init();
+        
         return pds;
     }
+    
+    
+    /**
+     * This reads in the (maven filtered) datasource properties from the test
+     * resource directory.
+     * 
+     * @return Properties containing the datasource properties.
+     */
+    private static Properties getDatasourceProperties() { 
+        boolean propertiesNotFound = false;
+        
+        // Central place to set additional H2 properties
+        System.setProperty("h2.lobInDatabase", "true");
+        
+        InputStream propsInputStream = HumanTaskServicesBaseTest.class.getResourceAsStream(DATASOURCE_PROPERTIES);
+        Properties props = new Properties();
+        if (propsInputStream != null) {
+            try {
+                props.load(propsInputStream);
+            } catch (IOException ioe) {
+                propertiesNotFound = true;
+                logger.warn("Unable to find properties, using default H2 properties: " + ioe.getMessage());
+                ioe.printStackTrace();
+            }
+        } else {
+            propertiesNotFound = true;
+        }
+
+        String password = props.getProperty("password");
+        if ("${maven.jdbc.password}".equals(password) || propertiesNotFound) {
+           logger.warn( "Unable to load datasource properties [" + DATASOURCE_PROPERTIES + "]" );
+        }
+        
+        // If maven filtering somehow doesn't work the way it should.. 
+        setDefaultProperties(props);
+
+        return props;
+    }
+
+    /**
+     * Return the default database/datasource properties - These properties use
+     * an in-memory H2 database
+     * 
+     * This is used when the developer is somehow running the tests but
+     * bypassing the maven filtering that's been turned on in the pom.
+     * 
+     * @return Properties containing the default properties
+     */
+    private static void setDefaultProperties(Properties props) {
+        String[] keyArr = { 
+                "serverName", "portNumber", "databaseName", JDBC_URL,
+                USER, PASSWORD,
+                DRIVER_CLASS_NAME, DATASOURCE_CLASS_NAME,
+                MAX_POOL_SIZE, ALLOW_LOCAL_TXS };
+        String[] defaultPropArr = { 
+                "", "", "", "jdbc:h2:mem:jbpm-db;MVCC=true",
+                "sa", "", 
+                "org.h2.Driver", "bitronix.tm.resource.jdbc.lrc.LrcXADataSource", 
+                "5", "true" };
+        Assert.assertTrue("Unequal number of keys for default properties", keyArr.length == defaultPropArr.length);
+        for (int i = 0; i < keyArr.length; ++i) {
+            if( ! props.containsKey(keyArr[i]) ) {
+                props.put(keyArr[i], defaultPropArr[i]);
+            }
+        }
+    }
+
 }
